requirements.txt
firecrawl-py
openai
python-dotenv 